% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{max.phenology}
\alias{max.phenology}
\title{Get the latest day of the year from a phenology object}
\usage{
\method{max}{phenology}(obj)
}
\arguments{
\item{obj}{A phenology object.}
}
\value{
The latest day of the year (1 to 365) of observations.
}
\description{
Get the latest day of the year from a phenology object
}
