% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{min.phenology}
\alias{min.phenology}
\title{Get the earliest day of the year from a phenology object}
\usage{
\method{min}{phenology}(obj)
}
\arguments{
\item{obj}{A phenology object.}
}
\value{
The earliest day of the year (1 to 365) of observations.
}
\description{
Get the earliest day of the year from a phenology object
}
