% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_clean.R
\name{pheno_clean}
\alias{pheno_clean}
\title{Clean Phenology Data by Removing Outliers}
\usage{
pheno_clean(phenology)
}
\arguments{
\item{phenology}{A phenology object containing occurrence data with day_of_year and latitude.}
}
\value{
A cleaned phenology object with outliers removed.
}
\description{
This function takes a phenology object, removes the top and bottom 5% outliers
based on the day of the year, removes outliers based on residuals
from a Generalized Additive Model (GAM) fitted to the latitude vs. day of the year,
and removes occurrences with day_of_year outliers within species groups.
}
\examples{
\dontrun{
# Assuming 'pheno' is a valid phenology object
cleaned_pheno <- clean_pheno(pheno)
}
}
