% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_load.R
\name{pheno_load}
\alias{pheno_load}
\title{Load Phenology Data from CSV}
\usage{
pheno_load(filename)
}
\arguments{
\item{filename}{The path to the CSV file containing iNaturalist occurrences.}
}
\value{
A phenology object containing:
  \item{name}{The scientific name of the species.}
  \item{occurrences}{A data table of occurrences with selected columns: `latitude`, `longitude`, `time_observed_at`, `positional_accuracy`.}
  \item{median_day}{Median day of the year (1 to 365) of observations.}
  \item{yearly_counts}{A list indicating the number of occurrences for each year represented.}
}
\description{
This function reads a CSV file and creates a phenology object.
}
