% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_overlap.R
\name{pheno_overlap}
\alias{pheno_overlap}
\title{Check for Phenological Overlap Between Two Species}
\usage{
pheno_overlap(species1_data, species2_data, lat_range = NULL, mode = "binary")
}
\arguments{
\item{species1_data}{A data frame containing phenology data for the first species with columns 'day_of_year' and 'latitude'.}

\item{species2_data}{A data frame containing phenology data for the second species with columns 'day_of_year' and 'latitude'.}

\item{lat_range}{A numeric vector of length 2 specifying the latitude range (c(lat_min, lat_max)) to restrict the analysis. Default is NULL, meaning the full data is used.}

\item{mode}{A string specifying the mode of overlap calculation. 'binary' (default) checks for any overlap in the day of year range. 'gam' uses Generalized Additive Models (GAM) to calculate overlap.}
}
\value{
A logical value or a numeric overlap area depending on the mode. Returns TRUE/FALSE in binary mode and a numeric value in gam mode.
}
\description{
This function determines if two species have any overlap in their range of days of year. By default, it uses the full phenology object. It can also evaluate if there is any phenological overlap specifically within a specified latitude range.
}
\examples{
species1_data <- data.frame(
  day_of_year = sample(100:200, 500, replace = TRUE),
  latitude = runif(500, min = 30, max = 60)
)
species2_data <- data.frame(
  day_of_year = sample(150:250, 500, replace = TRUE),
  latitude = runif(500, min = 30, max = 60)
)
# Check for overlap in binary mode
pheno_overlap(species1_data, species2_data)
# Check for overlap in gam mode
pheno_overlap(species1_data, species2_data, mode = "gam")
# Check for overlap with specified latitude range in binary mode
pheno_overlap(species1_data, species2_data, lat_range = c(40, 50))
# Check for overlap with specified latitude range in gam mode
pheno_overlap(species1_data, species2_data, lat_range = c(40, 50), mode = "gam")
}
