% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoccurrences.R
\name{phenoccurrences}
\alias{phenoccurrences}
\title{Get Occurrences of a Species from iNaturalist given a phenophase}
\usage{
phenoccurrences(
  species,
  spatial_bounds = c(15, -125, 65, -60),
  year_bounds = c(2015, 2024),
  phenophase = "flowering",
  max_results = 10000,
  min_results = 3,
  write = NULL
)
}
\arguments{
\item{species}{Character string specifying the species name to query.}

\item{spatial_bounds}{Numeric vector of length 4 specifying the spatial bounds for the query in the form \code{c(latitude_min, longitude_min, latitude_max, longitude_max)}. Default is \code{c(15, -125, 65, -60)}.}

\item{year_bounds}{Numeric vector of length 2 specifying the year bounds for the query. Default is \code{c(2015, 2024)}.}

\item{phenophase}{Character string specifying the phenophase to filter occurrences. Must be one of \code{"flowering"}, \code{"fruiting"}, \code{"budding"}, or \code{"none"}. Default is \code{"flowering"}.}

\item{max_results}{Integer specifying the maximum number of results to retrieve. Default is 10000.}

\item{min_results}{Integer specifying the minimum number of results required. If fewer results are found, the function will stop with an error. Default is 3.}

\item{write}{Character string specifying the path to write the results to a CSV file. If \code{NULL}, the results are not written to a file. Default is \code{NULL}.}
}
\value{
A \code{data.table} containing the occurrences of the specified species, with an additional column for the year of observation.
}
\description{
This function retrieves occurrences of a specified species from iNaturalist within defined spatial and temporal bounds, filtering by phenophase and other criteria.
}
\details{
The function queries the iNaturalist database for occurrences of the specified species within the given spatial and temporal bounds, filtered by the specified phenophase. The results are filtered to include only research-grade, georeferenced observations. If the number of occurrences found is less than \code{min_results}, the function stops with an error. If a file path is provided to \code{write}, the results are written to a CSV file.
}
\examples{
\dontrun{
# Get occurrences of "Abutilon theophrasti" flowering from 2015 to 2024
occurrences <- get_occurrences("Abutilon theophrasti")

# Get occurrences of "Abutilon theophrasti" fruiting from 2015 to 2024 and write to a file
occurrences <- get_occurrences("Abutilon theophrasti", phenophase="fruiting", write="occurrences.csv")
}

}
