% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{phenology}
\alias{phenology}
\title{Create a Phenology Object}
\usage{
phenology(data)
}
\arguments{
\item{data}{A data frame containing iNaturalist occurrences with columns: `scientific_name`, `latitude`, `longitude`, `time_observed_at`, `positional_accuracy`.}
}
\value{
A phenology object containing:
  \item{name}{The scientific name of the species.}
  \item{occurrences}{A data table of occurrences with selected columns: `latitude`, `longitude`, `time_observed_at`, `positional_accuracy`.}
  \item{median_day}{Median day of the year (1 to 365) of observations.}
  \item{yearly_counts}{A list indicating the number of occurrences for each year represented.}
}
\description{
This function creates a phenology object from a data frame of iNaturalist occurrences.
}
