% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{print.phenology}
\alias{print.phenology}
\title{Print a Phenology Object}
\usage{
\method{print}{phenology}(obj)
}
\arguments{
\item{obj}{A phenology object.}
}
\description{
Print method for objects of class `phenology`.
}
