% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{[.phenology}
\alias{[.phenology}
\title{Subset a Phenology Object}
\usage{
\method{[}{phenology}(obj, subset_expr)
}
\arguments{
\item{obj}{A phenology object.}

\item{subset_expr}{An expression to subset the occurrences data table within the phenology object.}
}
\value{
A phenology object updated with only occurrences matching the subset expression.
}
\description{
Subset method for objects of class `phenology`.
}
\details{
This method allows subsetting of a phenology object using standard bracket notation.
}
