% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{summary.phenology}
\alias{summary.phenology}
\title{Summary of a Phenology Object}
\usage{
\method{summary}{phenology}(obj)
}
\arguments{
\item{obj}{A phenology object.}
}
\description{
Summary method for objects of class `phenology`.
}
